+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - For plugin >= v10.4.0 use Homebridge UI >= v5.13.0
 - after update to v10.0.0 and above the accessory and bridge need to be removed from the homebridge / Home.app and added again
 
+## [10.5.1] - (02.03.2026)
+
+## Changes
+
+- added energy consumpion from P.V. today/lifetime statistics
+- fix power and energy statistics data
+- cleanup
+
 ## [10.5.0] - (01.03.2026)
 
 ## Changes